src/retriever/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "This is the container for the machine learning pipeline portion of the project"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.1.1",
    "einops>=0.8.1",
    "fastapi>=0.118.0",
    "google-cloud-storage>=3.4.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.1",
    "uvicorn>=0.37.0",
]
